refactoring. models updated. Get DetailListing view on listing click.

+++ Carfax/app/src/main/java/com/appsdeviser/carfax/data/model/SearchArea.kt
@@ -2,11 +2,11 @@ package com.appsdeviser.carfax.data.model
 
 data class SearchArea(
     val city: String,
-    val dynamicRadii: List<Int>,
+    val dynamicRadii: List<Double>,
     val dynamicRadius: Boolean,
     val latitude: Double,
     val longitude: Double,
-    val radius: Int,
+    val radius: Double,
     val state: String,
     val zip: String
 )
